==> 1. Dataset. 2. News month files to Dataset.ipynb <==
Берем все полученные файлы, убираем HTML и возвращаем датасет

import os
import sys 
from bs4 import BeautifulSoup
import json
import re
import datetime
import pandas as pd
import openpyxl

abs_path = os.path.abspath(os.getcwd()) + '\\'
dir_path = abs_path + 'finam_news_dataset'
dir_path

загружаем файлы новостей по месяцам, и объединяем

data = []
for path in os.listdir(dir_path):
  path = dir_path + '\\' + path
  with open(path, 'r', encoding='utf-8') as f:
    temp = json.load(f)
    for item in temp:
      data.append(item)

def replaceTextContent(s):
  s = s.encode().decode().replace('\t', '').replace('\xa0', ' ')
  s = re.sub(r'(\n[ ]*)+', '\n', s)
  return s

def parseHTML(item):
  soup = BeautifulSoup(item[3])
  item[3] = replaceTextContent(soup.get_text())
  return item[1:]

data_copy = data.copy()
sys.getsizeof(data_copy)

data_copy=list(map(lambda item: parseHTML(item), data_copy))

len(data_copy)

sys.getsizeof(data_copy)

print(data_copy[10000])

print(data_copy[-10000])

print(data_copy[-1000])
data_copy[-1000][2]

def comparator(item):
  return datetime.datetime.strptime(item, '%Y-%m-%d %H:%M') 

data_copy = sorted(data_copy, key=lambda item: comparator(item[1]))


# Сохранение в XLSX

import xlsxwriter

workbook = xlsxwriter.Workbook('news.xlsx')
worksheet = workbook.add_worksheet()

col = 0
worksheet.write_row(0, col, ['title','datetime','news_body'])
for row, data in enumerate(data_copy):
    worksheet.write_row(row+1, col, data)

workbook.close()

df = pd.read_excel('news.xlsx')

df.to_csv('news.csv', index=False)

df_csv = pd.read_csv('news.csv')

df

df_csv
