==> 2. Preprocessing. 2. Stacking News By Previous Days.ipynb <==
Преобразуем датает, объединеняя новости за одну дату в одну строку

import nltk
import openpyxl
import pymorphy3
import json
import numpy as np
import pandas as pd
import datetime

df = pd.read_csv('news_processed.csv')

df

def getDateOnly(date):
  return datetime.date(date.year, date.month, date.day)

def getDayNews(df):
  result = pd.DataFrame()
  prev_date = getDateOnly(datetime.datetime.strptime(df.iloc[0]['datetime'], '%Y-%m-%d %H:%M'))
  day_news = ''

  n = len(df)
  for i in range(n):
    
    curr_date = getDateOnly(datetime.datetime.strptime(df.iloc[i]['datetime'], '%Y-%m-%d %H:%M'))
    if (prev_date == curr_date):
      day_news += str(df.iloc[i]['news_body']) + " "
    else: 
      result = pd.concat([result, pd.DataFrame([{'date':curr_date, 'news_body': day_news}])])
      prev_date = curr_date
      day_news = str(df.iloc[i]['news_body']) + " "
  result = result.reset_index(drop=True)
  return result

df.iloc[4]['news_body']

df_by_days = getDayNews(df)

df_by_days

сохранение новостей объединенных по дням

df_by_days.to_csv('news_by_days.csv',index=False)

## Собираем датасет за несколько предыдущих дней

df_by_days = pd.read_csv('news_by_days.csv')

def getCummulative(df, days):
  result = pd.DataFrame()
  n =  len(df)
  cummulative = list(df['news_body'][:days])

  prev_date = getDateOnly(datetime.datetime.strptime(df.iloc[days-1]['date'], '%Y-%m-%d'))
  for i in range(days, n):
    curr_date = getDateOnly(datetime.datetime.strptime(df.iloc[i]['date'], '%Y-%m-%d'))

    # если даты пропущены то берем, что есть
    if (curr_date - prev_date).days > 1:
      temp_date = prev_date
      while (curr_date - temp_date).days > 1:
        temp_date += datetime.timedelta(days=1)
        result = pd.concat([result, pd.DataFrame([{'date':temp_date, 'news_body': ''.join(cummulative)}])])

    result = pd.concat([result, pd.DataFrame([{'date':curr_date, 'news_body': ''.join(cummulative)}])])
    
    # убираем первый элемент и добавляем в конец текущий
    cummulative.pop(0)
    cummulative.append(df.iloc[i]['news_body'])
    prev_date = curr_date

  result = result.reset_index(drop=True)
  return result

сохранение новостей с несколькими предыдущими датами, если даты пропущены, то берутся значение окна с предыдущей даты

for i in [1,3,5,7,10,14]:
  getCummulative(df_by_days, i).to_csv(f'news_by_days_{i}.csv', index=False)

df_result = pd.read_csv('news_by_days_1.csv')

def getDateDiffs(dates):
  dates.values.tolist()
  result = []
  for i in range(1, len(dates)):
    result.append((getDateOnly(dates[i]) - getDateOnly(dates[i-1])).days)
  df = pd.DataFrame(result, columns=['DATE_DIFF'])
  return df.groupby(['DATE_DIFF'])['DATE_DIFF'].count()

getDateDiffs(df_result['date'].apply(lambda x: getDateOnly(datetime.datetime.strptime(x, '%Y-%m-%d')) ))

df_result.iloc[2500:2510]
