==> 2. Preprocessing. 3. Generate Y.ipynb <==
import nltk
import openpyxl
import pymorphy3
import json
import numpy as np
import pandas as pd
import datetime
import sklearn

import datetime

def string_date_parse(data):
  return datetime.datetime.strptime(str(data), '%Y%m%d').date()

market_directory = 'market_data'
market_files = {
    'sber':'SBER.csv',
    'lukoil':'LKOH.csv',
    'gazprom':'GAZP.csv',
    'moex_i':'MOEX.csv'
}

market_data = {}

for key in market_files.keys():
  temp = pd.read_csv(f"{market_directory}/{market_files[key]}")
  # удаляем ненужные символы из названия колонок
  temp.columns = [elem.replace('<','').replace('>','') for elem in temp.columns]
  # удаляем ненужные столбцы
  temp = temp.drop(['PER', 'TIME'], axis=1)

  # среднее значение между максимумом и минимумом
  mid = temp.apply(lambda data: (data['HIGH'] + data['LOW'])/2, axis=1)
  # вставляем столбец со средним значением
  temp.insert(list(temp.columns).index('HIGH')+1, 'MID', mid)

  # преобразуем дату к datetime
  temp['DATE'] = temp['DATE'].apply(lambda data: string_date_parse(data))

  market_data[key] = temp

market_data['moex_i']

def getDateOnly(date):
  return datetime.date(date.year, date.month, date.day)

import math
def interpolate_data(data_start, data_end):
  ticker = data_start['TICKER']
  current_date = data_start['DATE']
  days_number = (getDateOnly(data_end['DATE'])-getDateOnly(data_start['DATE'])).days
  data_delta = data_start.copy()

  del data_delta['DATE']
  del data_delta['TICKER']

  for key in data_delta.keys():
    data_delta[key] = (data_end[key] - data_start[key])/days_number

  df = []
  for i in range(1, days_number):
    temp = data_delta.copy()

    for key in temp.keys():
      temp[key] = data_start[key] + temp[key]*i

    temp['VOL'] = int(math.floor(temp['VOL']))
    temp['DATE'] = (getDateOnly(data_start['DATE']) + datetime.timedelta(days=i))
    temp['TICKER'] = ticker

    df.append(temp)

  return df

Интерполируем недостающие даты: значения будут равны средним значениям

def market_interpolate(df):
  n = len(df)
  i = 1
  while i < n:
    if (getDateOnly(df.iloc[i]['DATE']) - getDateOnly(df.iloc[i - 1]['DATE'])).days > 1:
      new_data = interpolate_data(df.iloc[i - 1], df.iloc[i])
      interpolated = pd.DataFrame(new_data)
      df = pd.concat([df.iloc[:i], interpolated, df.iloc[i:]]).reset_index(drop=True)
      n = len(df)
    i += 1

  return df

date_default = market_data['sber']['DATE']
date_default_set = {x for x in market_data['sber']['DATE']}

for key in market_data.keys():
  market_data[key] = market_interpolate(market_data[key])

date_converted = market_data['sber']['DATE']
date_converted_set = {x for x in market_data['sber']['DATE']}
len(date_converted)

date_converted_set.difference(date_default_set)

## Финальные предсказания 

def getYWithDate(stocks):
  y = {'DATE':[], 'Y':[]}
  n = len(stocks)
  for i in range(1, n):
    data = stocks.loc[i]
    prev_data = stocks.loc[i-1]
    percent_delta = (data['MID'] - prev_data['MID'])/prev_data['MID']
    
    y['DATE'].append(data['DATE'])
    y['Y'].append(percent_delta)
  return y  

market_data['moex_i']

sberY = pd.DataFrame(getYWithDate(market_data['moex_i']),columns=['DATE', 'Y'])
sberY

def getDateDiffs(dates):
  dates.values.tolist()
  result = []
  for i in range(1, len(dates)):
    result.append((getDateOnly(dates[i]) - getDateOnly(dates[i-1])).days)
  df = pd.DataFrame(result, columns=['DATE_DIFF'])
  return df.groupby(['DATE_DIFF'])['DATE_DIFF'].count()

смотрим пропущенные дни. Число больше единицы означает пропуск в n-1 день

getDateDiffs(date_default)

getDateDiffs(date_converted)

# Сохраняем предсказания для всех существующих ценных бумаг

pd.DataFrame(getYWithDate(market_data['sber']),columns=['DATE', 'Y'])

market_data.keys()

сохранение предсказаний Y в виде процета изменения стоимости по отношению к предыдущей дате

for key in market_data.keys():
  pd.DataFrame(getYWithDate(market_data[key]),columns=['DATE', 'Y']).to_csv(f'./market_data/{key}_Y.csv', index=False)

сохранение полного датасета

for key in market_data.keys():
  pd.DataFrame(market_data[key]).to_csv(f'./market_data/{key}_FULL.csv', index=False)
