==> 3. Analysis. 2. Stocks.ipynb <==
import nltk
import openpyxl
import pymorphy3
import json
import numpy as np
import pandas as pd
import datetime
import sklearn

def getDateOnly(date):
  return datetime.date(date.year, date.month, date.day)
def getDateOnlyFromString(date_string):
  return getDateOnly(datetime.datetime.strptime(str(date_string), '%Y-%m-%d').date())
def getDateDiffs(dates):
  dates.values.tolist()
  result = []
  for i in range(1, len(dates)):
    result.append((getDateOnly(dates[i]) - getDateOnly(dates[i-1])).days)
  df = pd.DataFrame(result, columns=['DATE_DIFF'])
  return df.groupby(['DATE_DIFF'])['DATE_DIFF'].count()

market_path = 'market_data/sber_FULL.csv'

sber = pd.read_csv(market_path) 
sber['DATE'] = sber['DATE'].apply(lambda x: getDateOnlyFromString(x))
sber

sber_Y = pd.read_csv('market_data/sber_Y.csv')
sber_Y

sber_Y['Y'] = sber_Y['Y'].apply(lambda x: x>0)

#sber_Y['Y'] = sber_Y['Y'].apply(lambda x: x*100)

sber_no_Date = sber.drop(['DATE', 'TICKER'], axis=1).drop(index=[0]).reset_index(drop=True)
sber_no_Date

start = 4000
end = 6000

dataset_X = sber_no_Date.iloc[start:end]
dataset_Y = sber_Y['Y'].iloc[start:end]

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(dataset_X, 
                                                    dataset_Y, 
                                                    test_size=0.2, 
                                                    random_state=777)

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC, SVR
from xgboost import XGBClassifier
from sklearn import metrics

models = [LogisticRegression(), SVR(), SVC(), XGBClassifier()]
 
for i in range(3):
  models[i].fit(X_train, y_train)

  print(f'{models[i]} : ')
  print('Training Accuracy : ', metrics.roc_auc_score(
    y_train, models[i].predict(X_train)))
  print('Validation Accuracy : ', metrics.roc_auc_score(
    y_test, models[i].predict(X_test)))
  print()

from keras import backend as K

def recall_m(y_true, y_pred):
    true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
    possible_positives = K.sum(K.round(K.clip(y_true, 0, 1)))
    recall = true_positives / (possible_positives + K.epsilon())
    return recall

def precision_m(y_true, y_pred):
    true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
    predicted_positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
    precision = true_positives / (predicted_positives + K.epsilon())
    return precision

def f1_m(y_true, y_pred):
    precision = precision_m(y_true, y_pred)
    recall = recall_m(y_true, y_pred)
    return 2*((precision*recall)/(precision+recall+K.epsilon()))

import tensorflow as tf

from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from tensorflow.keras.layers import Conv1D, MaxPooling1D
from tensorflow.keras.layers import Flatten
from tensorflow.keras.layers import Embedding
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.layers import Dense, Dropout, Activation, Input, Reshape
from tensorflow.keras.layers import LSTM
from keras.layers import LeakyReLU

X_train.shape

# from keras.layers import LeakyReLU
# model = Sequential()
# model.add(Dense(6, kernel_initializer ='glorot_uniform',input_dim=X_train.shape[1]))
# model.add(LeakyReLU(alpha=0.01))
# model.add(Dropout(0.20))
# model.add(Dense(8, kernel_initializer ='glorot_uniform'))
# model.add(LeakyReLU(alpha=0.01))
# model.add(Dropout(0.20))
# # model.add(LSTM(4))
# model.add(Dropout(0.20))
# model.add(Dense(4))

# model.add(Dense(1, kernel_initializer ='glorot_uniform', activation = 'sigmoid'))
# model.compile(loss='binary_crossentropy',
#               optimizer='adamax',
#               metrics=['accuracy',f1_m,precision_m, recall_m])

from keras.layers import LeakyReLU
model = Sequential()
#model.add(Dense(6,input_dim=X_train.shape[1], activation='relu'))
model.add(Input(shape=(X_train.shape[1],)))
#model.add(Reshape((1,X_train.shape[1])))
#model.add(LSTM(2, input_shape=(1, X_train.shape[1]), return_sequences=False))
model.add(Dense(6, activation='relu'))
model.add(Dense(1, activation = 'sigmoid'))
model.compile(loss='binary_crossentropy',
              optimizer='adam',
              metrics=['accuracy',f1_m,precision_m, recall_m])
es = tf.keras.callbacks.EarlyStopping(monitor='val_f1_m', min_delta=0, patience=3, verbose=0, mode='auto', baseline=None, restore_best_weights=False)

# model = Sequential()
# model.add(Dense(6, kernel_initializer ='glorot_uniform',input_dim=X_train.shape[1]))
# model.add(LeakyReLU(alpha=0.01))
# model.add(Dropout(0.20))
# model.add(Dense(1, kernel_initializer ='glorot_uniform'))
# model.add(LeakyReLU(alpha=0.01))
# model.add(Dropout(0.20))
# model.add(Dense(1, kernel_initializer ='glorot_uniform', activation = 'sigmoid'))
# model.compile(loss='binary_crossentropy',
#               optimizer='adamax',
#               metrics=['acc',f1_m,precision_m, recall_m])

#es = tf.keras.callbacks.EarlyStopping(monitor='val_f1_m', min_delta=0, patience=3, verbose=0, mode='auto', baseline=None, restore_best_weights=False)

model.summary()

model.fit(X_train, y_train, batch_size = 4, epochs = 40, callbacks=[es])

y_pred = model.predict(X_test)

y_pred = np.array([i>0.5 for i in y_pred.flatten()])
y_pred

from sklearn.metrics import f1_score, precision_score,recall_score, precision_recall_curve

f1_score(y_test, y_pred)

precision_score(y_test, y_pred)

recall_score(y_test, y_pred)
